flask-server/requirements.txt
-i https://pypi.org/simple
bcrypt==4.1.3; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
flask==3.0.3
flask-bcrypt==1.0.1
flask-cors==4.0.1
flask-jwt-extended==4.6.0
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
pymysql==1.1.1
python-dotenv==1.0.1
werkzeug==3.0.3; python_version >= '3.8'
